receive ioctl should check unsupported features
When receiving a send stream that has unexpected features enabled (i.e.
a flag is set in `DMU_GET_FEATUREFLAGS(drr_versioninfo)` that is not in
`DMU_BACKUP_FEATURE_MASK`), libzfs will detect this and fail, printing
`cannot receive: stream has unsupported feature, feature flags = ...`.

However, if a process is using libzfs_core directly (i.e.
`lzc_receive*()`), or if libzfs and the kernel module are mismatched,
then the kernel may attempt to receive a stream with unsupported
features.

The problem is that the kernel is not explicitly checking
drr_versioninfo for unsupported features.  Typically this will cause the
receive the fail with EINVAL when it encounters a specific piece of
metadata that is not understood (e.g. a new record type).  However,
other, arbitrarily bad failure modes are also possible.

The solution is to make the kernel explicitly check drr_versioninfo for
unsupported features.  A new zfs_errno_t value was added, which the
kernel will return, and libzfs will print `cannot receive new filesystem
stream: the loaded zfs module does not support this operation. A reboot
may be required to enable this operation.`