ZFS header package should put files in one directory
Previously we generated the files for the header package by copying
what's in `include` directory of the source tree.  Since the source file
restructuring for ZoF, the header files in the source code were
reorganized into platform-specific subdirectories (for Linux and
FreeBSD).  However, when installed, we want all the header files
together, as they were before the ZoF restructuring.

`make install` knows how to arrange the header files from the
platform-specific subdirectories into a single combined directory.  This
change leverages that to generate the directory structure for the header
package.  Specifically, we `make install -C include` into a temporary
directory, and then use that temporary directory to get the files for
the package.